AudioFrame._recalc_linesize is less drastic
This fixes pops we were having in resampled audio.
We need to recalculate the line size after
resampling, but the function we were using was
also resetting the data pointers based off of the
new line size, despite the data be filled by the
resampler based off the old line size.