Add support for more ndarray <-> VideoFrame conversions
We support round-tripping:

- 24-bit RGB
- 32-bit RGBA (new)
- 8-bit grayscale (new)
- yuv420p (new)
- yuyv422 (new)

We drop the gray16 conversion code as it was not handling byte order
correctly, and could not be triggered anyway due to an impossible test:

  frame.format.name == ('gray16le', 'gray16be')

We also support creating VideoFrame with a user-specified alignment.
This is required in order to test the conversions work correctly for
frames with an alignment other than 1. Such frames are often output by
decoders for optimization reasons.

Merges #415.