Use av_image_alloc to allocate image buffer
It enforces alignment constraints on plane pointers, which keeps image
manipulations sane. We also ensure that the output buffer used by
sws_scale is at least 8-byte aligned, or it bombs due to out of bounds
writes.

Ideally, we should enforce a minimum alignment of 8-bytes for all
VideoFrames we create, but we need to fix to_image and from_image first.