Assert PyAV is built against a compatible FFmpeg (fixes: #401)
We explicitly fail the build as early as possible if a build is
attempted against an unsupported FFmpeg version. This gives users a
clear indication of what is wrong instead of triggering obscure
compiler errors about missing functions or struct members.

We also remove the reference to PYAV_HAVE_FFMPEG which is not defined
anymore.