Ensure `close` is called when Container is deleted
Currently, close() is only called when using a context manager. This can
demonstrably lead to hangs when using AUTO threading mode as
shown in the cookbook example.

As @jlaine discovered in #909, calling `close()` on the Container
seems to resolve the issue, likely by making sure the threads have
been joined by the time __dealloc__ runs.