Restore tests against FFmpeg 5.x
This reverts e884e6cd1dac2646fd002b441ca4fdfb2ef818c5.

So far we have not made any changes which justify dropping FFmpeg 5.x,
and we do not change the supported platforms in a minor release. We will
drop compatibility for FFmpeg 5.x once we switch to the new audio
channel API.