Fix `Buffer` subclass type definitions
It does not make sense to special-case the `Packet` class concerning
buffer protocol support, we have multiple classes inhering our `Buffer`
class:

- `Packet`
- `Plane` and its subclasses `AudioPlane` and `VideoPlane`
- `SideData`

We cannot use `collection.abc.Buffer`, yet as this is only available in
Python 3.12 and up. We do however want to declare that `bytes(foo)` is
available, so declare a `__bytes__` method in `Buffer`.