Make VideoFrame.from_numpy_buffer support buffers with padding
Some devices have hardware that creates image buffers with padding, so
adding support here means less frame buffer copying is required.

Specifically, we extend the support to buffers where the pixel rows
are contiguous, though the image doesn't comprise all the pixels on
the row (and is therefore not strictly contiguous).

We also support yuv420p images with padding. These have padding in the
middle of the UV rows as well as at the end, so can't be trimmed by
the application before being passed in. Instead, the true image width
must be passed.

Tests are also added to ensure all these cases now avoid copying.