Add typing overloads for `CodecContext.create`
The return type of `CodecContext.create` can be narrowed based on the
name of the requested codec, allowing us to know whether we are dealing
with an `AudioCodecContext` or a `VideoCodecContext`.

As the same logic is used by `OutputContainer.add_stream`, use a shared
list of audio / video codec names. We add the following codec names:

- audio : libopus, pcm_alaw, pcm_mulaw
- video : libvpx, libx264